lint migrations and make sure they are dont fail if they have been already executed

+++ Database/Migrations/2020_04_28_160649_OptimizeHasUPG0506PdfCareplanMediaQuery.php
@@ -1,27 +1,15 @@
 <?php
 
-use Illuminate\Support\Facades\Schema;
-use Illuminate\Database\Schema\Blueprint;
+/*
+ * This file is part of CarePlan Manager by CircleLink Health.
+ */
+
 use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
 
 class OptimizeHasUPG0506PdfCareplanMediaQuery extends Migration
 {
-    /**
-     * Run the migrations.
-     *
-     * @return void
-     */
-    public function up()
-    {
-        Schema::table('media', function (Blueprint $table){
-            $table->boolean('is_pdf')->virtualAs( 'IF(JSON_UNQUOTE(custom_properties->"$.is_pdf") = "true", TRUE, FALSE)')->before('created_at')->index();
-            $table->boolean('is_ccda')->virtualAs( 'IF(JSON_UNQUOTE(custom_properties->"$.is_ccda") = "true", TRUE, FALSE)')->before('created_at')->index();
-            $table->boolean('is_upg0506')->virtualAs( 'IF(JSON_UNQUOTE(custom_properties->"$.is_upg0506") = "true", TRUE, FALSE)')->before('created_at')->index();
-            $table->boolean('is_upg0506_complete')->virtualAs( 'IF(JSON_UNQUOTE(custom_properties->"$.is_upg0506_complete") = "true", TRUE, FALSE)')->before('created_at')->index();
-            $table->string('mrn')->virtualAs( 'JSON_UNQUOTE(custom_properties->"$.care_plan.demographics.mrn_number")')->before('created_at')->index();
-        });
-    }
-
     /**
      * Reverse the migrations.
      *
@@ -29,12 +17,36 @@ class OptimizeHasUPG0506PdfCareplanMediaQuery extends Migration
      */
     public function down()
     {
-        Schema::table('media', function (Blueprint $table){
+        if ( ! Schema::hasColumn('media', 'is_pdf')) {
+            return;
+        }
+
+        Schema::table('media', function (Blueprint $table) {
             $table->dropColumn('is_pdf');
             $table->dropColumn('is_ccda');
             $table->dropColumn('is_upg0506');
             $table->dropColumn('is_upg0506_complete');
             $table->dropColumn('mrn');
+        });
+    }
+
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        if (Schema::hasColumn('media', 'is_pdf')) {
+            return;
+        }
+
+        Schema::table('media', function (Blueprint $table) {
+            $table->boolean('is_pdf')->virtualAs('IF(JSON_UNQUOTE(custom_properties->"$.is_pdf") = "true", TRUE, FALSE)')->before('created_at')->index();
+            $table->boolean('is_ccda')->virtualAs('IF(JSON_UNQUOTE(custom_properties->"$.is_ccda") = "true", TRUE, FALSE)')->before('created_at')->index();
+            $table->boolean('is_upg0506')->virtualAs('IF(JSON_UNQUOTE(custom_properties->"$.is_upg0506") = "true", TRUE, FALSE)')->before('created_at')->index();
+            $table->boolean('is_upg0506_complete')->virtualAs('IF(JSON_UNQUOTE(custom_properties->"$.is_upg0506_complete") = "true", TRUE, FALSE)')->before('created_at')->index();
+            $table->string('mrn')->virtualAs('JSON_UNQUOTE(custom_properties->"$.care_plan.demographics.mrn_number")')->before('created_at')->index();
         });
     }
 }
